move matrix from algebra to math lib
