+++ libs/math/test/matrix.cpp
@@ -29,16 +29,17 @@
 #include <boost/test/data/test_case.hpp>
 #include <boost/test/data/monomorphic.hpp>
 
-#include <nil/crypto3/algebra/matrix/matrix.hpp>
-#include <nil/crypto3/algebra/matrix/math.hpp>
-#include <nil/crypto3/algebra/matrix/operators.hpp>
-#include <nil/crypto3/algebra/matrix/utility.hpp>
-#include <nil/crypto3/algebra/vector/vector.hpp>
-#include <nil/crypto3/algebra/vector/operators.hpp>
+#include <nil/crypto3/math/matrix/matrix.hpp>
+#include <nil/crypto3/math/matrix/math.hpp>
+#include <nil/crypto3/math/matrix/operators.hpp>
+#include <nil/crypto3/math/matrix/utility.hpp>
+#include <nil/crypto3/math/vector/vector.hpp>
+#include <nil/crypto3/math/vector/operators.hpp>
 
 #include <nil/crypto3/algebra/fields/goldilocks.hpp>
 
 using namespace nil::crypto3::algebra;
+using namespace nil::crypto3::math;
 
 using field = fields::goldilocks;
 using value = field::value_type;
